Improve security tests and add coverage attributes
- Rename and improve XXE test for clarity per code review
- Add new test for LIBXML_NONET network blocking
- Add CoversNothing attribute to all security tests (they don't cover specific classes)
- Add better documentation in test methods

+++ src/Api/Response.php
@@ -97,7 +97,7 @@ class Response
      */
     public static function paginated(array $data, int $total, int $page, int $perPage): void
     {
-        $totalPages = ceil($total / $perPage);
+        $totalPages = (int)ceil($total / $perPage);
         
         self::success([
             'items' => $data,
